Update deployment spec & add role spec
Update the partial deployment spec to not set ATOMIST_POD_NAMESPACE,
since that is no longer necessary as it can be read from the file in
the pod, remove properties that have sensible defaults, and add
security contexts and related volumes/volumeMounts.

Add partial role spec that has rules necessary to manage launching and
monitoring jobs.